requirements.txt
fastapi
uvicorn
llama-cpp-python
aiofiles
python-multipart
